+++ src/main/kotlin/de/visualdigits/kotlin/minim/audio/AudioStream.kt
@@ -20,7 +20,7 @@ interface AudioStream : AudioResource {
      * The size of the returned array will be the same size as getFormat().getChannels().
      */
     @Deprecated("")
-    fun read(): DoubleArray
+    fun read(): FloatArray
 
     /**
      * Reads buffer.getBufferSize() sample frames and puts them into buffer's channels.
